+++ python/cxx/eclipse_config.cpp
@@ -4,6 +4,10 @@
 #include <opm/parser/eclipse/EclipseState/InitConfig/InitConfig.hpp>
 #include <opm/parser/eclipse/EclipseState/SummaryConfig/SummaryConfig.hpp>
 #include <opm/parser/eclipse/EclipseState/SimulationConfig/SimulationConfig.hpp>
+#include <opm/parser/eclipse/EclipseState/Tables/TableManager.hpp>
+#include <opm/parser/eclipse/EclipseState/EclipseState.hpp>
+#include <opm/parser/eclipse/EclipseState/Schedule/Schedule.hpp>
+
 
 #include "export.hpp"
 
@@ -15,6 +19,10 @@ void python::common::export_EclipseConfig(py::module& module)
       .def( "restart",         &EclipseConfig::restart, ref_internal);
 
     py::class_< SummaryConfig >( module, "SummaryConfig")
+        .def(py::init([](const Deck& deck, const EclipseState& state, const Schedule& schedule) {
+           return SummaryConfig( deck, schedule, state.getTableManager() );
+         }  )  )
+
         .def( "__contains__",    &SummaryConfig::hasKeyword );
 
     py::class_< InitConfig >( module, "InitConfig")
