--- python/python/opm/io/config/__init__.py
@@ -1,1 +1,0 @@
-from .config import *
