+++ python/python/opm/io/schedule/__init__.py
@@ -1,3 +1,1 @@
 from opm._common import Schedule
-
-from .schedule import *
