+++ python/python/opm/io/summary/__init__.py
@@ -1,0 +1,1 @@
+from opm._common import SummaryConfig
